+++ src/rogallo/commands/ui.py
@@ -41,13 +41,6 @@ class ToggleSidePanel(Command):
 
 
 ##############################################################################
-class ToggleLinkNumbers(Command):
-    """Toggle the display of link numbers in the document viewer"""
-
-    BINDING_KEY = "shift+f8"
-
-
-##############################################################################
 class ToggleCosyLinkNumbers(Command):
     """Toggle the position of link numbers when they're being displayed"""